Djano-LangGragh-Agent/requirements.txt
Django>=5.0,<6.0
jupyter
langchain-openai
langgraph
langgraph-supervisor
python-decouple
permit
djangorestframework